Removes "remote" cli flag
The remote cli flag may not be useful if the repository will be cloned
into a temp directory, as it'll always be "origin".  Left the plumbing
in for now, in case it's more productive to act on the existing
checked-out repository in the cwd, instead.  Unsure of the best
workflow at the moment, and need feedback.

Signed-off-by: Christopher Collins <collins.christopher@gmail.com>